requirements.txt
requests==2.25.1
youtube-transcript-api==0.4.4
google-api-python-client==2.31.0
pydantic==1.8.2


superagi-tools
requests
pydantic
